Extract context middleware
